+++ CHANGELOG.md
@@ -57,6 +57,8 @@ In order to build the DUNE core modules you need at least the following software
 - Add the utility `IsCompileTimeConstant` to check for integral constants and anything with
   the same interface.
 
+- Add the macro `DUNE_ASSUME` for portable compiler assumption.
+
 ## Build System
 
 - Documentation files in `doc/buildsystem/${ModuleName}.rst` are now only copied.